adding generate info method
