+++ ddf_library/functions/etl/parallelize.py
@@ -4,10 +4,10 @@
 from pycompss.api.task import task
 from pycompss.functions.reduce import merge_reduce
 from pycompss.api.api import compss_wait_on, compss_delete_object
-from ddf_library.utils import generate_info, merge_schema
+from ddf_library.utils import merge_schema, _get_schema
 
-import math
 import pandas as pd
+import numpy as np
 import sys
 
 
@@ -59,7 +59,7 @@ def _generate_distribution1(n_rows, nfrag):
     in fragments at end and that doesnt have more available data.
     """
     size = n_rows / nfrag
-    size = int(math.ceil(size))
+    size = int(np.ceil(size))
     sizes = [size for _ in range(nfrag)]
 
     return sizes
@@ -69,7 +69,7 @@ def _generate_distribution2(n_rows, nfrag):
     """Data is splitted among the partitions."""
 
     size = n_rows / nfrag
-    size = int(math.ceil(size))
+    size = int(np.ceil(size))
     sizes = [size for _ in range(nfrag)]
 
     i = 0
@@ -112,7 +112,11 @@ def _check_schema(info):
 
     memory = info['memory']
     human_readable = _human_bytes(sum(memory))
-    print("Size of pandas in memory (bytes): ", memory)
+    avg = _human_bytes(np.mean(memory))
+    median = _human_bytes(np.median(memory))
+
+    print("Memory size of a "
+          "partition: avg({}), median({})".format(avg, median))
     print("Total size of pandas in memory: ", human_readable)
 
     return info
@@ -132,8 +136,4 @@ def _human_bytes(size):
     return value
 
 
-@task(returns=1)
-def _get_schema(df, f):
-    info = generate_info(df, f)
-    return info
