+++ ddf_library/functions/statistics/kolmogorov_smirnov.py
@@ -5,8 +5,8 @@ __author__ = "Lucas Miguel S Ponce"
 __email__ = "lucasmsp@gmail.com"
 
 from ddf_library.utils import merge_info
-from ..etl.select import select
-from ..etl.sort import SortOperation
+from ddf_library.functions.etl.select import select
+from ddf_library.functions.etl.sort import SortOperation
 
 from pycompss.api.task import task
 from pycompss.functions.reduce import merge_reduce
@@ -18,62 +18,65 @@ from scipy.stats import distributions
 
 
 def kolmogorov_smirnov_one_sample(data, settings):
-        """
-        Perform the Kolmogorov-Smirnov test for goodness of fit. This
-        implementation of Kolmogorov–Smirnov test is a two-sided test
-        for the null hypothesis that the sample is drawn from a continuous
-        distribution.
+    """
+    Perform the Kolmogorov-Smirnov test for goodness of fit. This
+    implementation of Kolmogorov–Smirnov test is a two-sided test
+    for the null hypothesis that the sample is drawn from a continuous
+    distribution.
 
-        :param data: A list with of pandas's DataFrame;
-        :param settings: A dictionary that contains:
-            - col: sample column name;
-            - distribution: Name of distribution (default is 'norm');
-            - args: A tuple of distribution parameters;
-            - mode: Defines the distribution used for calculating the p-value,
-             'approx' to use approximation to exact distribution or 'asymp' to
-             use asymptotic distribution of test statistic
-        :return: KS statistic and p-value
+    :param data: A list with of pandas's DataFrame;
+    :param settings: A dictionary that contains:
+        - col: sample column name;
+        - distribution: Name of distribution (default is 'norm');
+        - args: A tuple of distribution parameters;
+        - mode: Defines the distribution used for calculating the p-value,
+         'approx' to use approximation to exact distribution or 'asymp' to
+         use asymptotic distribution of test statistic
+    :return: KS statistic and p-value
 
-        .. seealso:: Visit this `link <https://docs.scipy.org/doc/scipy-0.14.0/
-         reference/stats.html#module-scipy.stats>`__ to see all supported
-         distributions.
+    .. seealso:: Visit this `link <https://docs.scipy.org/doc/scipy-0.14.0/
+     reference/stats.html#module-scipy.stats>`__ to see all supported
+     distributions.
 
-        """
+    """
 
-        nfrag = len(data)
-        col = settings['col']
-        mode = settings.get('mode', 'asymp')
+    nfrag = len(data)
+    col = settings['col']
+    mode = settings.get('mode', 'asymp')
 
-        if mode not in ['asymp', 'asymp']:
-            raise Exception("Only supported `approx` or `asymp` mode.")
+    if mode not in ['asymp', 'asymp']:
+        raise Exception("Only supported `approx` or `asymp` mode.")
 
-        if not isinstance(col, list):
-            col = [col]
+    if not isinstance(col, list):
+        col = [col]
 
-        info = [0 for _ in range(nfrag)]
-        for f in range(nfrag):
-            params = {'columns': col, 'id_frag':f}
-            data[f], info[f] = _select(data[f], params)
-        info = merge_info(info)
-        info = compss_wait_on(info)
+    info = [0 for _ in range(nfrag)]
+    for f in range(nfrag):
+        params = {'columns': col, 'id_frag': f}
+        data[f], info[f] = _select(data[f], params)
+    info = merge_info(info)
+    info = compss_wait_on(info)
 
-        params = {
-            'columns': col,
-            'ascending': [True for _ in col],
-            'info': [info]
-        }
+    params = {
+        'columns': col,
+        'ascending': [True for _ in col],
+        'info': [info]
+    }
 
-        sort_output = SortOperation().transform(data, params)
-        sort_data = sort_output['data']
-        compss_delete_object(data)
+    # range_partition used in sort operation can create less partitions
+    sort_output = SortOperation().transform(data, params)
+    sort_data, info = sort_output['data'], sort_output['info']
+    info = merge_info(info)
+    nfrag = len(sort_data)
 
-        for f in range(nfrag):
-            sort_data[f] = _ks_theorical_dist(sort_data[f], info, f, settings)
-        info = merge_reduce(_ks_merge, sort_data)
-        compss_delete_object(sort_data)
+    for f in range(nfrag):
+        sort_data[f] = _ks_theorical_dist(sort_data[f], info, f, settings)
 
-        info = _ks_d_critical(info, mode)
-        return info
+    info = merge_reduce(_ks_merge, sort_data)
+    compss_delete_object(sort_data)
+
+    info = _ks_d_critical(info, mode)
+    return info
 
 
 @task(returns=2)
