+++ tests/benchmark/std_scaler/test.py
@@ -1,42 +1,74 @@
-#!/usr/bin/python
+#!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
-from pycompss.api.task import task
 from ddf_library.ddf import DDF
-from ddf_library.functions.ml.feature import VectorAssembler, StandardScaler
+from ddf_library.utils import generate_info
+from ddf_library.functions.ml.feature import StandardScaler, VectorAssembler
+
+from pycompss.api.api import compss_barrier
+from pycompss.api.task import task
 
 import pandas as pd
 import numpy as np
+import time
 
 
-@task(returns=1)
-def generate_partition(size, col_name, dim):
-    df = pd.DataFrame()
-    df[col_name] = np.random.random_integers(1, 1000, size=(size, dim)).tolist()
-    return df
+@task(returns=2)
+def generate_partition(size, cols, dim, frag):
+
+    df = pd.DataFrame(np.random.randint(1, 1000*size, size=(size, dim)),
+                      columns=cols)
+    info = generate_info(df, frag)
+    return df, info
 
 
 def generate_data(total_size, nfrag, col_name, dim):
 
-    size = total_size / nfrag
+    dfs = [[] for _ in range(nfrag)]
+    info = [[] for _ in range(nfrag)]
+
+    size = total_size // nfrag
     sizes = [size for _ in range(nfrag)]
     sizes[-1] += (total_size - sum(sizes))
 
-    dfs = [generate_partition(size, col_name, dim) for size in sizes]
+    for f, s in enumerate(sizes):
+        dfs[f], info[f] = generate_partition(s, col_name, dim, f)
 
-    return dfs
+    return dfs, info
 
 
 if __name__ == "__main__":
-    print "\n|-------- STD Scaler --------|\n"
-    total_size = int(sys.argv[1])
-    nfrag = int(sys.argv[2])
+
+    n_rows = int(sys.argv[1])
+    n_frag = int(sys.argv[2])
     dim = 2
+    cols = ["col{}".format(c) for c in range(dim)]
     col_name = 'features'
 
-    df_list = generate_data(total_size, nfrag, col_name, dim)
+    t1 = time.time()
+    df_list, info = generate_data(n_rows, n_frag, cols, dim)
+    ddf1 = DDF().import_data(df_list, info)
+    ddf1 = VectorAssembler(input_col=cols, output_col=col_name, remove=True)\
+        .transform(ddf1).cache()
 
-    ddf1 = DDF().import_data(df_list)
+    compss_barrier()
+    t2 = time.time()
+    print("t2-t1:", t2 - t1)
 
-    ddf1 = StandardScaler(input_col=col_name).fit_transform(ddf1).cache()
-    # print ddf1.show()
+    ddf_train, ddf_test = ddf1.split(0.7)
+    ddf_train = ddf_train.cache()
+    ddf_test = ddf_test.cache()
+
+    compss_barrier()
+    t3 = time.time()
+    print("t3-t1:", t3 - t1)
+
+    scaler = StandardScaler(input_col=col_name).fit(ddf_train)
+    ddf1 = scaler.transform(ddf1)
+    compss_barrier()
+    t4 = time.time()
+    print("t4-t3:", t4 - t3)
+    print("t_all:", t4 - t1)
+
+    # print(ddf_train.count())
+    # print(ddf_test.count())
