+++ tests/test_ml_classifiers.py
@@ -1,5 +1,6 @@
-#!/usr/bin/python
+#!/usr/bin/env python3
 # -*- coding: utf-8 -*-
+
 
 from ddf_library.ddf import DDF
 import numpy as np
