+++ tests/test_ml_clustering.py
@@ -1,41 +1,44 @@
-#!/usr/bin/python
+#!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
-from ddf.ddf import DDF
+from ddf_library.ddf import DDF
 import pandas as pd
 
-
-def ml_clustering():
-
+def base():
     df = pd.DataFrame([[1.0, 2.0], [1.0, 4.0], [1.0, 0], [4.0, 2.0],
                        [4.0, 4.0], [4.0, 0.0]], columns=['x', 'y'])
 
     # creating DDF from a DataFrame
     ddf = DDF().parallelize(df, 4)
 
-    from ddf.functions.ml.feature import VectorAssembler
+    from ddf_library.functions.ml.feature import VectorAssembler
     assembler = VectorAssembler(input_col=["x", "y"], output_col="features")
     ddf = assembler.transform(ddf).drop(['x', 'y'])
 
     # scaling features using MinMax Scaler
-    from ddf.functions.ml.feature import MinMaxScaler
-    scaler = MinMaxScaler(input_col='features', output_col='features_norm')\
+    from ddf_library.functions.ml.feature import MinMaxScaler
+    scaler = MinMaxScaler(input_col='features', output_col='features_norm') \
         .fit(ddf)
     ddf = scaler.transform(ddf)
 
-    from ddf.functions.ml.clustering import Kmeans
+    return ddf
+
+
+def kmeans(ddf):
+
+    from ddf_library.functions.ml.clustering import Kmeans
     kmeans = Kmeans(feature_col='features_norm', n_clusters=2,
                     init_mode='random', pred_col='kmeans1').fit(ddf)
     kmeans.save_model('/kmeans')
-    ddf = kmeans.transform(ddf)
+
+    del kmeans
 
     # to test save and load models
     kmeans = Kmeans(feature_col='features_norm', n_clusters=2,
                     init_mode='random', pred_col='kmeans2')\
         .load_model('/kmeans')
-    df = kmeans.transform(ddf).show(15)
+    kmeans.transform(ddf).show(15)
 
-    print "Kmeans result:\n", df
     """
                features features_norm  kmeans1  kmeans2
         0  [1.0, 2.0]    [0.0, 0.5]      0.0      0.0
@@ -48,5 +51,6 @@ def ml_clustering():
 
 
 if __name__ == '__main__':
-    print "_____Testing Machine Learning Clustering_____"
-    ml_clustering()
+    print("_____Testing Machine Learning Clustering_____")
+    ddf_test = base()
+    kmeans(ddf_test)
